+++ app/src/androidTest/java/com/averycorp/prismtask/CloudIdOrphanHealerTwoDeviceTest.kt
@@ -342,6 +342,11 @@ class CloudIdOrphanHealerTwoDeviceTest {
                     weeklyReviewDao = db.weeklyReviewDao(),
                     dailyEssentialSlotCompletionDao = db.dailyEssentialSlotCompletionDao(),
                     attachmentDao = db.attachmentDao(),
+                    medicationDao = db.medicationDao(),
+                    medicationDoseDao = db.medicationDoseDao(),
+                    medicationSlotDao = db.medicationSlotDao(),
+                    medicationSlotOverrideDao = db.medicationSlotOverrideDao(),
+                    medicationTierStateDao = db.medicationTierStateDao(),
                     logger = logger
                 )
                 return Device(db, healer, firestore, label)